+++ plastron-legacy/src/plastron/cli.py
@@ -115,7 +115,7 @@ def main():
         config = envsubst(yaml.safe_load(args.config_file))
         repo_config = config['REPOSITORY']
         broker_config = config.get('MESSAGE_BROKER', None)
-        command_config = config.get('COMMANDS', {})
+        command_config = config.get('COMMANDS') or {}
     else:
         # old-style, repository-only config file
         with open(args.repo, 'r') as repo_config_file:
